requirements.txt
fastapi-redis-cache
fastapi==0.73.0
motor==2.5.1
pydantic[email]
python-decouple==3.6
uvicorn==0.17.4

PyJWT
bcrypt
passlib
locust
pytest
httpx
SQLAlchemy==1.4.26
sqlalchemy2-stubs
gunicorn==20.1.0